ComplainApi/requirements.txt
python-dotenv
langchain
pyproject-toml
langchain_groq
langchain_community
pillow
transformers
torchvision
torch
flask
flask-cors
flask-caching
firebase-admin
pandas
tabulate
langchain_google_genai
faiss-cpu